Try moving the assertions file
