+++ analytics/tests/assertions.py
@@ -1,0 +1,10 @@
+"""Utilities for common assertion patterns."""
+
+
+def assert_dict_contains(d: dict, expected: dict) -> None:
+    """
+    Assert that d contains all the key-value pairs in expected.
+
+    Do this by checking to see if adding `expected` to `d` leaves `d` unchanged.
+    """
+    assert d | expected == d
